+++ HISTORY.rst
@@ -36,6 +36,9 @@
 
 Unreleased Changes
 ------------------
+* Workbench
+    * Fixed a bug where the workbench could crash if there was too much
+      std error from an invest model.
 * Urban Cooling
     * Updated the text for the ``building_intensity`` column in the biophysical
       table to clarify that the values of this column should be normalized